--- Reciplease/Reciplease/Model/CoreData/Favorite 2.swift
@@ -1,15 +1,0 @@
-//
-//  Favorite.swift
-//  Reciplease
-//
-//  Created by Bertrand Dalleau on 28/06/2022.
-//
-
-import Foundation
-import SharkORM
-
-class Favorite: SRKObject {
-
-    @objc dynamic var recipe: Recipe?
-
-}
