--- Reciplease/Reciplease/View/IngredientTableViewCell 2.swift
@@ -1,25 +1,0 @@
-//
-//  IngredientTableViewCell.swift
-//  Reciplease
-//
-//  Created by Bertrand Dalleau on 24/06/2022.
-//
-
-import UIKit
-
-class IngredientTableViewCell: UITableViewCell {
-
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-    @IBOutlet weak var ingredientLabel: UILabel!
-    
-}
